+++ include/libcommute/libcommute.hpp
@@ -25,6 +25,7 @@
 #include "expression/generator_fermion.hpp"
 #include "expression/generator_spin.hpp"
 #include "expression/hc.hpp"
+#include "loperator/compressed_state_view.hpp"
 #include "loperator/elementary_space_boson.hpp"
 #include "loperator/elementary_space_fermion.hpp"
 #include "loperator/elementary_space_spin.hpp"